Analyze Gust Factor

Analyzes the ratio of peak gust to sustained wind speed. This is the wind equivalent of the wave Hmax/Hs ratio.

Usage

analyze_gust_factor(data, min_wind_speed = 5)

Arguments

data

Data frame with gust and wind_speed columns

min_wind_speed

Minimum sustained wind speed to consider (default: 5 m/s)

Value

List with:

summary: summary statistics of gust factor

extreme_gusts: observations with high gust factors

by_wind_category: gust factor by wind speed category

Examples

if (FALSE) { # \dontrun{
con <- connect_duckdb()
data <- query_buoy_data(con, variables = c("time", "wind_speed", "gust"))
gust_analysis <- analyze_gust_factor(data)
DBI::dbDisconnect(con)
} # }