Fit GEV Distribution to Annual Maxima

Fits a Generalized Extreme Value distribution to annual maximum values. This is the Block Maxima approach to extreme value analysis.

Usage

fit_gev_annual_maxima(
  data,
  variable = "wave_height",
  time_col = "time",
  min_years = 5
)

Arguments

data

Data frame with columns: time, value (the variable to analyze)

variable

Name of the variable column (default: "wave_height")

time_col

Name of the time column (default: "time")

min_years

Minimum years of data required (default: 5)

Value

List with:

fit: extRemes fevd object

annual_maxima: data frame of annual maxima

parameters: GEV parameters (location, scale, shape)

diagnostics: model diagnostic information

Examples

if (FALSE) { # \dontrun{
con <- connect_duckdb()
data <- query_buoy_data(con, variables = c("time", "wave_height"))
gev_result <- fit_gev_annual_maxima(data)
DBI::dbDisconnect(con)
} # }